Report: refactor handelbar helpers into static methods
moved from static helper class to object based

addressed comments

Replacing everything to camel case

changed from es6 to function based syntax
