+++ lighthouse-core/formatters/partials/table.html
@@ -3,11 +3,11 @@
 {{#if rows}}
 <details class="subitem__details">
   <summary class="subitem__detail">More information</summary>
-  <table class="table_list {{#if_not_eq headings.length 2}}multicolumn{{/if_not_eq}}">
+  <table class="table_list {{#ifNotEq headings.length 2}}multicolumn{{/ifNotEq}}">
     <thead>
       <tr>
         {{#each ../tableHeadings}}
-          <th {{#if_eq @key "preview"}}class="preview-image"{{/if_eq}}>{{this}}</th>
+          <th {{#ifEq @key "preview"}}class="preview-image"{{/ifEq}}>{{this}}</th>
         {{/each}}
       </tr>
     </thead>
