+++ lighthouse-core/report/report-generator.js
@@ -20,168 +20,15 @@
 
 const Formatter = require('../formatters/formatter');
 const Handlebars = require('handlebars');
+const handlebarHelpers = require('./handlebar-helpers');
 const fs = require('fs');
 const path = require('path');
-const marked = require('marked');
 
-const RATINGS = {
-  GOOD: {label: 'good', minScore: 75},
-  AVERAGE: {label: 'average', minScore: 45},
-  POOR: {label: 'poor'}
-};
-
-function calculateRating(value) {
-  let rating = RATINGS.POOR.label;
-  if (value >= RATINGS.GOOD.minScore) {
-    rating = RATINGS.GOOD.label;
-  } else if (value >= RATINGS.AVERAGE.minScore) {
-    rating = RATINGS.AVERAGE.label;
-  }
-  return rating;
-}
 
 class ReportGenerator {
 
   constructor() {
-    const getTotalScore = aggregation => {
-      return Math.round(aggregation.total * 100);
-    };
-
-    const getItemRating = value => {
-      if (typeof value === 'boolean') {
-        return value ? RATINGS.GOOD.label : RATINGS.POOR.label;
-      }
-      return calculateRating(value);
-    };
-
-    // Converts a name to a link.
-    Handlebars.registerHelper('nameToLink', name => {
-      return name.toLowerCase().replace(/\s/g, '-');
-    });
-
-    // Figures out the total score for an aggregation
-    Handlebars.registerHelper('getTotalScore', getTotalScore);
-
-    // Converts the total score to a rating that can be used for styling.
-    Handlebars.registerHelper('getTotalScoreRating', aggregation => {
-      const totalScore = getTotalScore(aggregation);
-      return calculateRating(totalScore);
-    });
-
-    // Converts a value to a rating string, which can be used inside the report
-    // for color styling.
-    Handlebars.registerHelper('getItemRating', getItemRating);
-
-    Handlebars.registerHelper('shouldShowHelpText',
-      value => (getItemRating(value) !== RATINGS.GOOD.label));
-
-    // Convert numbers to fixed point decimals
-    Handlebars.registerHelper('decimal', number => {
-      if (number && number.toFixed) {
-        return number.toFixed(2);
-      }
-      return number;
-    });
-
-    // value is boolean?
-    Handlebars.registerHelper('is-bool', value => (typeof value === 'boolean'));
-
-    // !value
-    Handlebars.registerHelper('not', value => !value);
-
-    // value === value2?
-    Handlebars.registerHelper('if_eq', function(lhs, rhs, options) {
-      if (lhs === rhs) {
-        // eslint-disable-next-line no-invalid-this
-        return options.fn(this);
-      } else {
-        // eslint-disable-next-line no-invalid-this
-        return options.inverse(this);
-      }
-    });
-
-    // value !== value2
-    Handlebars.registerHelper('if_not_eq', function(lhs, rhs, options) {
-      if (lhs !== rhs) {
-        // eslint-disable-next-line no-invalid-this
-        return options.fn(this);
-      } else {
-        // eslint-disable-next-line no-invalid-this
-        return options.inverse(this);
-      }
-    });
-
-    // arg1 && arg2 && ... && argn
-    Handlebars.registerHelper('and', function(...args) {
-      let arg = false;
-      for (let i = 0, n = args.length - 1; i < n; i++) {
-        arg = args[i];
-        if (!arg) {
-          break;
-        }
-      }
-      return arg;
-    });
-
-    // eslint-disable-next-line no-unused-vars
-    Handlebars.registerHelper('sanitize', function(str, opts) {
-      // const isViewer = opts.data.root.reportContext === 'viewer';
-
-      // Allow the report to inject HTML, but sanitize it first.
-      // Viewer in particular, allows user's to upload JSON. To mitigate against
-      // XSS, define a renderer that only transforms links and code snippets.
-      // All other markdown ad HTML is ignored.
-      const renderer = new marked.Renderer();
-      renderer.em = str => `<em>${str}</em>`;
-      renderer.link = (href, title, text) => {
-        const titleAttr = title ? `title="${title}"` : '';
-        return `<a href="${href}" target="_blank" rel="noopener" ${titleAttr}>${text}</a>`;
-      };
-      renderer.codespan = function(str) {
-        return `<code>${str}</code>`;
-      };
-      // eslint-disable-next-line no-unused-vars
-      renderer.code = function(code, language) {
-        return `<pre>${code}</pre>`;
-      };
-      renderer.image = function(src, title, text) {
-        return `<img src="${src}" alt="${text}" title="${title}">`;
-      };
-
-      // Nuke wrapper <p> tag that gets generated.
-      renderer.paragraph = function(str) {
-        return str;
-      };
-
-      try {
-        str = marked(str, {renderer, sanitize: true});
-      } catch (e) {
-        // Ignore fatal errors from marked js.
-      }
-
-      // The input str has been sanitized and transformed. Mark it as safe so
-      // handlebars renders the text as HTML.
-      return new Handlebars.SafeString(str);
-    });
-
-    // format time
-    Handlebars.registerHelper('format_time', function(date) {
-      const options = {
-        day: 'numeric', month: 'numeric', year: 'numeric',
-        hour: 'numeric', minute: 'numeric', second: 'numeric',
-        timeZoneName: 'short'
-      };
-      let formatter = new Intl.DateTimeFormat('en-US', options);
-
-      // Force UTC if runtime timezone could not be detected.
-      // See https://github.com/GoogleChrome/lighthouse/issues/1056
-      const tz = formatter.resolvedOptions().timeZone;
-      if (!tz || tz.toLowerCase() === 'etc/unknown') {
-        options.timeZone = 'UTC';
-        formatter = new Intl.DateTimeFormat('en-US', options);
-      }
-      return formatter.format(new Date(date));
-    });
+    Handlebars.registerHelper(handlebarHelpers);
   }
 
   /**
