+++ lighthouse-core/report/templates/report-template.html
@@ -55,11 +55,11 @@ limitations under the License.
             {{#each reportsCatalog.reportsMetadata }}
               <a class="menu__report-tab" href="{{ reportHref }}">
                 <p class="menu__report-tab__url" title="{{ url }}">Report for: {{ url }}</p>
-                <p class="menu__report-tab__time">Generated on: {{ format_time generatedTime }}</p>
+                <p class="menu__report-tab__time">Generated on: {{ formatDateTime generatedTime }}</p>
               </a>
-              {{#if_eq reportHref ../reportsCatalog.selectedReportHref }}
+              {{#ifEq reportHref ../reportsCatalog.selectedReportHref }}
                 {{> createNav ../aggregations }}
-              {{/if_eq}}
+              {{/ifEq}}
             {{/each}}
           {{else}}
             {{> createNav aggregations }}
@@ -71,7 +71,7 @@ limitations under the License.
           <button class="report-body__header-toggle js-header-toggle" title="See report's runtime settings"></button>
           <div class="report-body__metadata">
             <div class="report-body__url">Results for: <a href="{{ url }}" target="_blank" rel="noopener">{{ url }}</a></div>
-            <div class="report-body__url">Generated on: {{ format_time generatedTime}}</div>
+            <div class="report-body__url">Generated on: {{ formatDateTime generatedTime}}</div>
           </div>
           <div class="report-body__buttons">
             <div class="export-section">
@@ -81,9 +81,9 @@ limitations under the License.
                 <li><a href="#" class="report__icon copy" data-action="copy">Copy JSON</a></li>
                 <li><a href="#" class="report__icon download" data-action="save-html" data-context="{{reportContext}}">Save as HTML</a></li>
                 <li><a href="#" class="report__icon download" data-action="save-json" data-context="{{reportContext}}">Save as JSON</a></li>
-                {{#if_not_eq reportContext "viewer" }}
+                {{#ifNotEq reportContext "viewer" }}
                   <li><a href="#" class="report__icon open" data-action="open-viewer" data-context="{{reportContext}}">Open in Viewer</a></li>
-                {{/if_not_eq}}
+                {{/ifNotEq}}
               </ul>
             </div>
             <button class="report__icon report-body__icon share js-share" title="Share report on GitHub"></button>
@@ -160,7 +160,7 @@ limitations under the License.
 
                     {{ sanitize subItem.description }}
 
-                    {{~#if (and subItem.displayValue (not (is-bool subItem.displayValue))) ~}}
+                    {{~#if (and subItem.displayValue (not (isBool subItem.displayValue))) ~}}
                       <strong class="subitem__raw-value">: {{ subItem.displayValue }}</strong>
                     {{/if}}
 
@@ -187,7 +187,7 @@ limitations under the License.
                     {{#if subItem.error}}
                       <span class="subitem-result__unknown score-warning-bg">N/A</span>
                     {{else}}
-                      {{#if (is-bool subItem.score)}}
+                      {{#if (isBool subItem.score)}}
                         {{#if subItem.score}}
                           <span class="subitem-result__good score-good-bg">Pass</span>
                         {{else}}
@@ -216,13 +216,13 @@ limitations under the License.
     </div>
 
     <footer class="footer">
-      Generated by <b>Lighthouse</b> {{lighthouseVersion}} on {{ format_time generatedTime }} | <a href="https://github.com/GoogleChrome/Lighthouse/issues" target="_blank" rel="noopener">File an issue</a>
+      Generated by <b>Lighthouse</b> {{lighthouseVersion}} on {{ formatDateTime generatedTime }} | <a href="https://github.com/GoogleChrome/Lighthouse/issues" target="_blank" rel="noopener">File an issue</a>
     </footer>
-    {{#if_eq reportContext "perf-x" }}
+    {{#ifEq reportContext "perf-x" }}
     <div class="report-body__fixed-footer-container">
       {{> config-panel}}
     </div>
-    {{/if_eq}}
+    {{/ifEq}}
   </section>
 </div>
