+++ lighthouse-core/test/report/report-generator-test.js
